+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-### **(V.7.9.0 Changes) (26.1.0 Minecraft)**
+TODO: DO NOT RELEASE UNTIL COMPLETED
+
+- Essence Block Shader is working
+- Make TradeHintParticle work again 
+
+
+
+
+
+
+
+### **(V.7.9.0 Changes) (26.1.2 Minecraft)**
 
 #### Misc:
-Ported to 26.1.0
+Ported to 26.1.2